requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1
pytest-html>=3.2.0
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
pytest-randomly>=3.12.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-flask>=1.2.0
pytest-sqlalchemy>=0.1.0
pytest-env>=1.1.0
pytest-cache>=1.0
pytest-json-report>=1.5.0
pytest-html>=3.2.0
pytest-clarity>=1.0.1
pytest-sugar>=0.9.7
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
pytest-randomly>=3.12.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
pytest-flask>=1.2.0
pytest-sqlalchemy>=0.1.0
pytest-env>=1.1.0
pytest-cache>=1.0
pytest-json-report>=1.5.0
freezegun>=1.2.2
responses>=0.23.1
factory-boy>=3.3.0
faker>=19.3.0
